HR_Recuiter_Whatsapp_Chatbot/requirements.txt
flask
python-dotenv
aiohttp
requests
langchain
langchain-community
python-dotenv
faiss-cpu
sentence-transformers